love transaction recieve

+++ app/src/main/java/com/dk/pden/model/Transaction.kt
@@ -12,7 +12,7 @@ data class Transaction(
 
         var to: String,
 
-        var amount: Int,
+        var amount: Long,
 
         var actvity: String
 ) {
